refactor: extract persistence-free model-api from opennms-model
Create core/opennms-model-api module containing persistence-free types
extracted from opennms-model to break the split-package deadlock between
opennms-model (javax.persistence) and model-jakarta (jakarta.persistence).

Moved to model-api:
- 17 enums/interfaces/value types (OnmsSeverity, ResourcePath, etc.)
- EventBuilder + NodeLabelChangedEventBuilder (sans entity methods)
- ResourceTypeUtils (sans entity-dependent getNodeFromResource methods)

Entity-dependent methods extracted to ResourceNodeUtils in opennms-model.
OnmsNode.setIpInterface() calls in both opennms-model and model-jakarta
replaced with primitive setters (setNodeid + setInterface).

This enables JPA daemons to exclude opennms-model from their runtime
classpath while retaining EventBuilder and ResourceTypeUtils via model-api,
preventing Hibernate 7 from scanning javax.persistence-annotated entities.
