+++ core/opennms-model-api/src/main/java/org/opennms/netmgt/model/ResourcePath.java
