+++ core/opennms-model-api/src/main/java/org/opennms/netmgt/model/outage/OutageSummary.java
