+++ features/newts/src/main/java/org/opennms/netmgt/measurements/impl/NewtsFetchStrategy.java
@@ -48,6 +48,7 @@ import org.opennms.netmgt.measurements.utils.Utils;
 import org.opennms.netmgt.model.OnmsNode;
 import org.opennms.netmgt.model.OnmsResource;
 import org.opennms.netmgt.model.ResourceId;
+import org.opennms.netmgt.model.ResourceNodeUtils;
 import org.opennms.netmgt.model.ResourceTypeUtils;
 import org.opennms.netmgt.model.RrdGraphAttribute;
 import org.opennms.newts.api.Context;
@@ -419,12 +420,12 @@ public class NewtsFetchStrategy implements MeasurementFetchStrategy {
     private OnmsNode getNode(final OnmsResource resource, final Source source) {
         OnmsNode node = null;
         try {
-            node = ResourceTypeUtils.getNodeFromResourceRoot(resource);
+            node = ResourceNodeUtils.getNodeFromResourceRoot(resource);
         } catch (final ObjectRetrievalFailureException e) {
         }
         if (node == null) {
             final OnmsResource otherResource = m_resourceDao.getResourceById(ResourceId.fromString(source.getResourceId()).getParent());
-            node = ResourceTypeUtils.getNodeFromResource(otherResource);
+            node = ResourceNodeUtils.getNodeFromResource(otherResource);
         }
         return node;
     }
