+++ opennms-model/src/main/java/org/opennms/netmgt/model/OnmsNode.java
@@ -1582,7 +1582,10 @@ public class OnmsNode extends OnmsEntity implements Serializable, Comparable<Onm
 
         if(oldPrimaryInterface != null && scannedPrimaryIf != null){
             EventBuilder bldr = new EventBuilder(EventConstants.PRIMARY_SNMP_INTERFACE_CHANGED_EVENT_UEI, "Provisiond");
-            bldr.setIpInterface(scannedPrimaryIf);
+            if (scannedPrimaryIf.getNode() != null) {
+                bldr.setNodeid(scannedPrimaryIf.getNode().getId().longValue());
+            }
+            bldr.setInterface(scannedPrimaryIf.getIpAddress());
             bldr.setService("SNMP");
             bldr.addParam(EventConstants.PARM_OLD_PRIMARY_SNMP_ADDRESS, InetAddressUtils.str(oldPrimaryInterface.getIpAddress()));
             bldr.addParam(EventConstants.PARM_NEW_PRIMARY_SNMP_ADDRESS, InetAddressUtils.str(scannedPrimaryIf.getIpAddress()));
